re-implement event poll with a persistent system and a weight system
In the past the event poll was manually created with the owner needing
to keep track of what was played in the past and what games are
suppose to be on the list, with this new system they will only need
to be added to a list.

Previously there was also a cooldown system where a game that was
played last week could not be on the poll on the next week, this hurts
games that are new and highly in demand at the time (e.g. friendslop)
by forcing diversity.

The new system gives every entry a weight value that defaults to 1.0 and
ranges from 0.6 to 1.0 that gets applied to all the votes for it.
If the game is really popular it will be still be played but this will
stop repetition by giving it a disadvantage.

If a game wins it gets gets 0.2 subtracted to its weight and if it loses
it gets 0.1 added so for every week the game needs 2 weeks to recoup
up to 4 weeks.

The poll will automatically start If its Monday or later but not
beyond Friday, even it already started an automatic poll this week it
won't start another.