optimize poll processing by using reaction handlers
since karma and boosts can be changed while a poll is active it
recalculates the entire poll every minute to make sure its accurate.

This optimizes the system to take 500ms on average instead of upwards to
N seconds where N is the amount of entries
(e.g. 10 entries = 10 or more seconds)